requirements.txt
absl-py==2.1.0
attrs==23.2.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
filelock==3.15.4
flatbuffers==24.3.25
fonttools==4.53.0
fsspec==2024.6.1
idna==3.7
jax==0.4.30
jaxlib==0.4.30
Jinja2==3.1.4
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.0
mediapipe==0.10.11
ml-dtypes==0.4.0
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt-einsum==3.3.0
packaging==24.1
pillow==10.3.0
protobuf==3.20.3
pyarrow==16.1.0
pycparser==2.22
pyparsing==3.1.2
python-dateutil==2.9.0.post0
requests==2.32.3
rerun-sdk==0.16.1
scipy==1.14.0
six==1.16.0
sounddevice==0.4.7
sympy==1.12.1
torch==2.3.1
tqdm==4.66.4
triton==2.3.1
typing_extensions==4.12.2
urllib3==2.2.2
